spi: bcm2835: Fix for shared interrupts
BCM2711 shares an interrupt betweem 5 SPI interfaces (0, 3, 4, 5 & 6).
Another interrupt is shared between SPI1, SPI2 and UART1, which also
affects BCM2835/6/7. Acting on an interrupt intended for another
interface ought to be harmless (although potentially inefficient),
but it can cause this driver to crash - presumably because some
critical state is not ready.

Add a test to the spi-bcm2835 interrupt service routine that interrupts
are enabled on this interface to avoid the crash and improve efficiency.

Suggested by GitHub user boe-pi.

See: https://github.com/raspberrypi/linux/issues/5048

Signed-off-by: Phil Elwell <phil@raspberrypi.com>