overlays: Rename unofficial "rpi-" overlays
There are a number of overlays - some very old - with names beginning
with "rpi-", even though they aren't for products made by Raspberry Pi.
Rename them to make their status clearer. The entries in the overlay
map will translate the old names to the new ones, with a warning message
that may be displayed in the GUI.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>