arm64: dts: Remove the Pi5 FIRMWARE_UPDATED guards
The Raspberry Pi firmware has assumed that top level #size-cells value
in dtb files is 1. As a result, the dts source files have had to use
32-bit sizes, making it awkward to declare memory regions of 4GB or
larger, requiring them to be split into chunks. This primarily affects
Pi 5, where the dts source has made use of conditional compilation to
choose either 64-bit or 32-bit sizes, based on the presence or absence
of the defined cpp symbol FIRMWARE_UPDATED.

As of EEPROM release pieeprom-2025-02-11, the firmware has read and made
use of the actual #size-cells value declared in the dtb, allowing the
use of 64-bit sizes. Remove the conditional sections, retaining the
64-bit size values.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>